feat(chat): move stickToBottom lib to ui components
