+++ packages/react-instantsearch/src/widgets/Chat.tsx
@@ -4,8 +4,6 @@ import { find } from 'instantsearch.js/es/lib/utils';
 import React, { createElement, Fragment } from 'react';
 import { useInstantSearch, useChat } from 'react-instantsearch-core';
 
-import { useStickToBottom } from '../ui/lib/useStickToBottom';
-
 import { createSearchIndexTool } from './chat/tools/SearchIndexTool';
 
 export { SearchIndexToolType } from './chat/tools/SearchIndexTool';
@@ -15,7 +13,6 @@ import type {
   ChatProps as ChatUiProps,
   RecommendComponentProps,
   RecordWithObjectID,
-  MutableRef,
 } from 'instantsearch-ui-components';
 import type {
   AddToolResultWithOutput,
@@ -105,12 +102,7 @@ export function Chat<
   const [maximized, setMaximized] = React.useState(false);
   const [input, setInput] = React.useState('');
   const [isClearing, setIsClearing] = React.useState(false);
-
-  const { scrollRef, contentRef, isAtBottom, scrollToBottom } =
-    useStickToBottom({
-      initial: 'instant',
-      resize: 'smooth',
-    });
+  const [isScrollAtBottom, setIsScrollAtBottom] = React.useState(true);
 
   const tools = React.useMemo(() => {
     const defaults = createDefaultTools(itemComponent, getSearchPageURL);
@@ -208,10 +200,8 @@ export function Chat<
         setIndexUiState,
         isClearing,
         onClearTransitionEnd: handleClearTransitionEnd,
-        scrollRef: scrollRef as unknown as MutableRef<HTMLDivElement>,
-        contentRef: contentRef as unknown as MutableRef<HTMLDivElement>,
-        isScrollAtBottom: isAtBottom,
-        onScrollToBottom: scrollToBottom,
+        isScrollAtBottom,
+        setIsScrollAtBottom,
         ...messagesProps,
       }}
       promptProps={{
